# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_user.Address:
    properties:
      city:
        type: string
      created_at:
        type: string
      detail:
        type: string
      district:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      phone:
        type: string
      province:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  internal_user.Preference:
    properties:
      allergens:
        type: string
      dietary_limits:
        type: string
      id:
        type: integer
      scene_prefs:
        type: string
      taste_prefs:
        type: string
      user_id:
        type: integer
    type: object
  internal_user.Profile:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: integer
      phone:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: AI-powered snack e-commerce platform with hybrid search, conversational
    agent, and Stripe payments.
  title: Kotoha API
  version: "1.0"
paths:
  /addresses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_user.Address'
            type: array
      security:
      - BearerAuth: []
      summary: List user addresses
      tags:
      - profile
    post:
      consumes:
      - application/json
      parameters:
      - description: Address data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.Address'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_user.Address'
      security:
      - BearerAuth: []
      summary: Create a new address
      tags:
      - profile
  /addresses/{id}:
    delete:
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an address
      tags:
      - profile
    put:
      consumes:
      - application/json
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated address data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.Address'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.Address'
      security:
      - BearerAuth: []
      summary: Update an address
      tags:
      - profile
...
  /preferences:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.Preference'
      security:
      - BearerAuth: []
      summary: Get user preferences
      tags:
      - profile
    put:
      consumes:
      - application/json
      parameters:
      - description: Preference data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.Preference'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.Preference'
      security:
      - BearerAuth: []
      summary: Update user preferences
      tags:
      - profile
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.Profile'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.Profile'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - profile
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"